requirements.txt
absl-py==0.11.0
argon2-cffi
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs
backcall==0.2.0
beautifulsoup4==4.9.3
bleach
brython==3.9.3
cachetools==4.1.1
certifi==2020.12.5
cffi
chardet==3.0.4
click==7.1.2
colorama
comtypes==1.1.7
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==2.0.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
import-ipynb==0.1.3
importlib-metadata
ipykernel
ipython
ipython-genutils==0.2.0
itsdangerous==2.0.0
jedi
Jinja2==3.0.0
joblib==0.17.0
Js2Py==0.70
jsonschema
jupyter-client
jupyter-core==4.6.3
jupyterlab-pygments
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==2.0.0
mistune
mock==4.0.2
nbclient
nbconvert
nbformat
nest-asyncio
nltk==3.5
notebook
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.2
parso==0.7.0
pdftotext==2.1.5
pickleshare
Pillow==8.0.1
pipwin==0.5.0
prometheus-client==0.8.0
prompt-toolkit
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pygame==2.0.0
Pygments
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent
pySmartDL==1.3.4
python-aiml==0.9.3
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.4
pywinpty==0.5.7
pyzmq==19.0.2
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tensorboard==1.13.1
tensorboard-plugin-wit==1.7.0
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tflearn==0.3.2
tornado==6.0.4
tqdm==4.51.0
traitlets
tzlocal==2.1
urllib3==1.25.11
wcwidth
webencodings==0.5.1
Werkzeug==2.0.0
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.12.1
zipp
